added SuccessionTimestep to definition of Cohort but did not cast the type correctly in constructors; attempting to fix it

+++ src/ICohort.cs
@@ -9,7 +9,7 @@ namespace Landis.Library.PnETCohorts
     public interface ICohort : Library.UniversalCohorts.ICohort
     {
         new CohortData Data { get; }
-        ushort SuccessionTimestep { get; }
+        byte SuccessionTimestep { get; }
         int AGBiomass { get; }
         int TotalBiomass { get; }
         bool IsLeafOn { get; }
